Move axios hook to grid component
